+++ packages/core/src/lib/settings-manager/modules/difficulty/index.ts
@@ -1,2 +1,1 @@
-export * from './coloring';
 export * from './difficulty';
