+++ packages/core/src/lib/settings-manager/settings-manager.spec.ts
@@ -1,9 +1,7 @@
 import { v4 as uuid, validate } from 'uuid';
-import {
-  DefaultSettingsHandlerImpl,
-  TimerOnGuessDifficulty,
-} from './default-settings-handler';
+import { DefaultSettingsHandlerImpl } from './default-settings-handler';
 import { AlternateColoring } from './modules/coloring';
+import { TimerOnGuessDifficulty } from './modules/difficulty/difficulty';
 import { SettingsManager } from './settings-manager';
 
 describe('settingsManager', () => {
